ui: refactor search pages
* Removes generic SearchLayout

* Moves search layout of each collection to its SearchPage

* INSPIR-2485
