--- ui/src/common/layouts/SearchLayout/SearchLayout.scss
@@ -1,3 +1,0 @@
-.__SearchLayout__ {
-  margin: 20px 0;
-}
