--- ui/src/common/layouts/SearchLayout/index.js
@@ -1,3 +1,0 @@
-import SearchLayout from './SearchLayout';
-
-export default SearchLayout;
